timestamp going in but it's 1970. so is datatype off. should it be long?

+++ src/breslin/serverside/battle/battle.h
@@ -28,8 +28,8 @@ Combatant* mAwayCombatant;
 
 // battle time
 long mBattleTime;
-long mBattleStartTime;
-long mBattleEndTime;
+double mBattleStartTime;
+double mBattleEndTime;
 long mBattleLengthTime;
